+++ src/js/api/requestHttp.js
@@ -1,5 +1,10 @@
-import { SUCCESS_MESSAGE } from '../constants/message';
-import request from '../utils/request';
+import {
+  ERROR_MESSAGE,
+  INVALID_MESSAGE,
+  SUCCESS_MESSAGE,
+} from '../constants/message.js';
+import ValidationError from '../error/ValidationError.js';
+import request from '../utils/request.js';
 
 const api = {
   create: async (params, snackbar, updateSubwayState) => {
@@ -55,6 +60,24 @@ const api = {
     }
 
     return true;
+  },
+
+  login: async (params) => {
+    const response = await request.post(params);
+
+    if (response.status === 400) {
+      throw new ValidationError(INVALID_MESSAGE.LOGIN.FAILED);
+    }
+
+    if (!response.ok) throw Error(response.message);
+
+    const { accessToken } = await response.json();
+    return accessToken;
+  },
+
+  signup: async (params) => {
+    const response = await request.post(params);
+    if (!response.ok) throw Error(response.message);
   },
 };
